refactor: remove custom offline sync
